Create new Constraint and Violations in o.a.a.core.data.constraints
