+++ core/src/main/java/org/apache/accumulo/core/data/constraints/Constraint.java
@@ -16,36 +16,47 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.core.constraints;
+package org.apache.accumulo.core.data.constraints;
 
 import java.util.List;
 
 import org.apache.accumulo.core.data.Mutation;
-import org.apache.accumulo.core.dataImpl.KeyExtent;
+import org.apache.accumulo.core.data.TabletId;
 import org.apache.accumulo.core.security.AuthorizationContainer;
 
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-
 /**
- * This class is replaced by {@link org.apache.accumulo.core.data.constraints.Constraint}
+ * Constraint objects are used to determine if mutations will be applied to a table.
  *
- * @deprecated since 2.1.0 Use {@link org.apache.accumulo.core.data.constraints.Constraint}
+ * <p>
+ * This interface expects implementers to return violation codes. The reason codes are returned
+ * instead of arbitrary strings to encourage conciseness. Conciseness is needed because violations
+ * are aggregated. If a user sends a batch of 10,000 mutations to Accumulo, only aggregated counts
+ * about which violations occurred are returned. If the constraint implementer were allowed to
+ * return arbitrary violation strings like the following:
+ *
+ * <p>
+ * Value "abc" is not a number<br>
+ * Value "vbg" is not a number
+ *
+ * <p>
+ * This would not aggregate very well, because the same violation is represented with two different
+ * strings.
+ *
+ * @since 2.1.0 Replaces interface in org.apache.accumulo.core.constraints package
  */
-@Deprecated(since = "2.1.0")
-@SuppressFBWarnings(value = "NM_SAME_SIMPLE_NAME_AS_INTERFACE",
-    justification = "Same name used for compatibility during deprecation cycle")
-public interface Constraint extends org.apache.accumulo.core.data.constraints.Constraint {
+public interface Constraint {
 
   /**
    * The environment within which a constraint exists.
    */
   interface Environment {
+
     /**
-     * Gets the key extent of the environment.
+     * Gets the tablet Id of the environment.
      *
-     * @return key extent
+     * @return TabletId
      */
-    KeyExtent getExtent();
+    TabletId getTablet();
 
     /**
      * Gets the user within the environment.
