+++ core/src/test/java/org/apache/accumulo/core/data/constraints/DefaultKeySizeConstraintTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.core.constraints;
+package org.apache.accumulo.core.data.constraints;
 
 import static org.junit.Assert.assertEquals;
