+++ core/src/test/java/org/apache/accumulo/core/security/VisibilityConstraintTest.java
@@ -29,10 +29,10 @@ import static org.junit.Assert.assertNull;
 import java.util.Arrays;
 import java.util.List;
 
-import org.apache.accumulo.core.constraints.Constraint.Environment;
-import org.apache.accumulo.core.constraints.VisibilityConstraint;
 import org.apache.accumulo.core.data.ArrayByteSequence;
 import org.apache.accumulo.core.data.Mutation;
+import org.apache.accumulo.core.data.constraints.Constraint.Environment;
+import org.apache.accumulo.core.data.constraints.VisibilityConstraint;
 import org.junit.Before;
 import org.junit.Test;
