+++ server/base/src/main/java/org/apache/accumulo/server/metadata/RootTabletMutatorImpl.java
@@ -24,7 +24,6 @@ import java.util.List;
 
 import org.apache.accumulo.core.data.Mutation;
 import org.apache.accumulo.core.data.TabletId;
-import org.apache.accumulo.core.dataImpl.KeyExtent;
 import org.apache.accumulo.core.dataImpl.TabletIdImpl;
 import org.apache.accumulo.core.metadata.RootTable;
 import org.apache.accumulo.core.metadata.schema.Ample;
@@ -47,12 +46,6 @@ public class RootTabletMutatorImpl extends TabletMutatorBase implements Ample.Ta
 
     RootEnv(ServerContext ctx) {
       this.ctx = ctx;
-    }
-
-    @Override
-    @SuppressWarnings("deprecation")
-    public KeyExtent getExtent() {
-      return RootTable.EXTENT;
     }
 
     @Override
