+++ test/src/main/java/org/apache/accumulo/test/functional/SlowConstraint.java
@@ -23,8 +23,8 @@ import static org.apache.accumulo.fate.util.UtilWaitThread.sleepUninterruptibly;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-import org.apache.accumulo.core.constraints.Constraint;
 import org.apache.accumulo.core.data.Mutation;
+import org.apache.accumulo.core.data.constraints.Constraint;
 
 public class SlowConstraint implements Constraint {
